requirements.txt
python-telegram-bot
pandas
configparser
openpyxl
playwright
aiohttp